+++ src/main/kotlin/com/github/mohamead/spiderlog/icons/SpiderlogIcons.kt
@@ -1,0 +1,10 @@
+package com.github.mohamead.spiderlog.icons
+
+import com.intellij.openapi.util.IconLoader
+
+object SpiderlogIcons {
+
+    @JvmField
+    val toolWindowIcon = IconLoader.getIcon("/icons/toolWindowIcon.svg", javaClass)
+
+}