+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 # spiderlog Changelog
 
-## [Unreleased]
+## [0.0.1] - 2022-04-14
 ### Added
+- Setting(s) for spiderlog
+- Operation(s) for scrolling to top and end of window
+- Operation(s) for reading, updating and clearing .log file (content) to / from window  
+- Window with toolbar 
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)